Add metadata parsing from json file
In ldmx-sw v2.0 and on, the use of mac files is discontinued.

We can still read the metadata parameters we used to set in the config file
from there, but the ones from the mac file will instead come from a json
dump of all settable input parameters.

I have added a function that can be run with the arg "test", but it needs to be
properly included in the overall metadata extraction structure by being called the
right place.